Introduced ModifyingWithMethodsProviderWrapper
